requirements.txt
fastapi[standard]
pydantic-settings
python-jose[cryptography]
sqlalchemy